[GONG-50] action 파일 분리: (getServerTime, getTimetable)
